assignment problem are solved

+++ J_Factorial.c
@@ -1,11 +1,11 @@
-#include<stdio.h>
+#include <stdio.h>
 long long int rec(int n)
 {
     if (n == 1)
     {
-       return 1;
+        return 1;
     }
-   int mult = rec(n-1);
+    long long int mult = rec(n - 1);
     return n*mult;
 }
 int main()
